Add Config partition support.
Required changes:

1. Change make-flash to use config.img to create config partition.
2. Change Makefile to pass config.img to make-flash
3. Delete test-cert
4. Create test-conf, a container that creates a sample config partition
   with test certificates, and serves as template for future partition
   configuration generator.

Note: Now that test-conf takes files from zededa-container, the latter
is not a special snowflake anymore, and should be moved to pkg like any
other container (moved to pkg/zedctr).

Note: In a perfect world we would use LinuxKit's own pkg/mount, we can't
because the FAT filesystem (nor squashfs) is not supported, and can
work with UUID/Lables only on ext4, btrfs, xfs.